+++ src/components/Footer.js
@@ -1,0 +1,10 @@
+import React from "react";
+import './footer.css';
+
+export default function Footer(){
+    return(
+        <div className="footer">
+        <p>2021 Fidenz Technology</p>
+      </div>
+    )
+}
