+++ src/components/footer.css
@@ -1,0 +1,9 @@
+.footer {
+    bottom: 0;
+    margin-top: 50px;
+   width: 100%;
+    background-color: #121111;
+    padding: 10px;
+    text-align: center;
+    
+  }
