+++ src/components/weatherDetailsCard.css
@@ -1,0 +1,248 @@
+.card-container {
+ 
+    display: grid;
+    grid-template-columns: repeat(2, 1fr);
+    gap: 20px;
+    justify-items: center;
+    cursor: pointer;
+  }
+
+  .container{
+ 
+    position: relative;
+    margin-top: 49px;
+    margin-left: 69px;
+    height: 368px;
+    width: 579px;
+    border-radius: 10px;
+    box-shadow: 2px 2px 1px rgba(0,0,0,0.2);
+  }
+  .container:nth-child(1) {
+    background-image: linear-gradient(rgba(255, 132, 0, 0.5), rgba(255, 204, 0, 0.5)), url('../assets/cloudbg.png');
+  }
+  .container:nth-child(2) {
+    background-image: linear-gradient(rgba(0, 42, 255, 0.5), rgba(0, 51, 255, 0.5)), url('../assets/cloudbg.png');
+  }
+
+  .content { 
+    width: 579px;
+    height: 218px;
+    border-radius: 8px 8px 0 0;
+    background-size: cover;
+    background-repeat: no-repeat;
+    background-position: center;
+    position: relative;
+    margin-top: 0;
+  }
+
+  .sun{
+    z-index: 1000;
+    font-size: 15px !important;
+  }
+  
+  .location{
+    position: absolute;
+    font-family: "Roboto", sans-serif;
+    font-weight: 100;
+    font-size: 20px;
+    width: 177px;
+    height: 26px;
+    left: 69px;  
+    top: 49px;
+  }
+
+  .dateAndTime{
+    position: absolute;
+    font-family: "Roboto", sans-serif;
+    font-weight: 30;
+    font-size: 12px;
+    width: 150px;
+    height: 16px;
+    left: 80px;  
+    top: 91px;
+  }
+  .description{
+    position: absolute;
+    font-family: "Roboto", sans-serif;
+    font-weight: 40;
+    font-size: 14px;
+    width: 103px;
+    height: 16px;
+    left: 132px;  
+    top: 149px;
+  }
+  .temp{
+    z-index: 1000;
+    width: 126px;
+    height: 48px;
+    position: absolute;
+    font-family: "Noto Sans", sans-serif;
+    font-size: 44px;
+    font-weight: 20;
+    right: 100px;
+    top: 48px;
+  }
+  .tempMax{
+    z-index: 1000;
+    width: 130px;
+    height: 40px;
+    position: absolute;
+    font-family: "Noto Sans", sans-serif;
+    font-size: 14px;
+    font-weight: 100;
+    right: 110px;
+    top: 150px;
+  }
+
+  .tempMin{
+    z-index: 1000;
+    width: 130px;
+    height: 40px;
+    position: absolute;
+    font-family: "Noto Sans", sans-serif;
+    font-size: 14px;
+    font-weight: 100;
+    right: 110px;
+    top: 110px;
+  }
+  .bottom{
+  width: 579px;
+  height: 151px;
+  border-radius: 8px;
+  background-color: #383b47;
+  display: grid;
+  justify-content: center;
+  grid-template-columns: repeat(3, 160px);
+  gap: 2px;
+  border-radius: 0 0 10px 10px;
+  position: relative;
+  font-family: "Noto Sans", sans-serif;
+  }
+  .bottom > div {
+    list-style: none;
+    text-align: left;
+    padding: 20px 4px;
+    font-size: 14px;
+    justify-content: center;
+    margin: 5px;
+    height: 50%;
+    bottom: 5px;
+    top: 5px;
+  }
+  .vertical {
+    border-left: 1px solid rgb(134, 134, 149);
+    height: 100px;
+    position:relative;
+    margin-left: -6px;
+  }
+
+  @media screen and (max-width: 768px) {
+    .card-container {
+    display: flex;
+    flex-wrap: wrap;
+    margin: 0 -10px;
+    gap: 20px;
+    cursor: pointer;
+    }
+    
+    .container {
+    position: relative;
+    height: 218px;
+    width: 100%;
+    border-radius: 10px;
+    box-shadow: 2px 2px 1px rgba(0, 0, 0, 0.2);
+    }
+    .container:nth-child(1) {
+    background-image: linear-gradient(rgba(255, 132, 0, 0.5), rgba(255, 204, 0, 0.5)), url('../assets/cloudbg.png');
+    }
+    .container:nth-child(2) {
+    background-image: linear-gradient(rgba(0, 42, 255, 0.5), rgba(0, 51, 255, 0.5)), url('../assets/cloudbg.png');
+    }
+    
+    .content {
+    width: 100%;
+    height: 100%;
+    border-radius: 8px;
+    background-size: cover;
+    background-repeat: no-repeat;
+    background-position: center;
+    }
+    
+    .sun {
+    z-index: 1000;
+    font-size: 15px !important;
+    }
+    
+    .location {
+    position: absolute;
+    font-family: "Roboto", sans-serif;
+    font-weight: 100;
+    font-size: 18px;
+    width: 100%;
+    height: 26px;
+    top: 5px;
+    }
+    
+    .dateAndTime {
+    position: absolute;
+    font-family: "Roboto", sans-serif;
+    font-weight: 30;
+    font-size: 12px;
+    width: 100%;
+    height: 16px;
+    top: 35px;
+    }
+    
+    .description {
+    position: absolute;
+    font-family: "Roboto", sans-serif;
+    font-weight: 40;
+    font-size: 14px;
+    width: 100%;
+    height: 16px;
+    top: 55px;
+    }
+    
+    .temp {
+    z-index: 1000;
+    width: 100%;
+    height: 48px;
+    position: absolute;
+    font-family: "Noto Sans", sans-serif;
+    font-size: 44px;
+    font-weight: 20;
+    bottom: 48px;
+    }
+    
+    .tempMax {
+    z-index: 1000;
+    width: 100%;
+    height: 40px;
+    position: absolute;
+    font-family: "Noto Sans", sans-serif;
+    font-size: 14px;
+    font-weight: 100;
+    bottom: 150px;
+    }
+    
+    .tempMin {
+    z-index: 1000;
+    width: 100%;
+    height: 40px;
+    position: absolute;
+    font-family: "Noto Sans", sans-serif;
+    font-size: 14px;
+    font-weight: 100;
+    bottom: 110px;
+    }
+    
+    .bottom {
+    width: 100%;
+    height: 100px;
+    border-radius: 0;
+    background-color: #383b47;
+    display: flex;
+    justify-content: center;
+    grid-template-columns: repeat;
+    }
+  }
