--- src/index.css
@@ -1,331 +1,0 @@
-*{
-  padding: 0;
-  margin: 0;
-  color: #FFF;
-}
-
-.logo {
-  display: flex;
-  position: relative;
-  align-items: center;
-  margin-top: 50px;
-  justify-content: center; 
-  width: 100%; 
-  height: 85px;
-}
-.app {
-  display: flex;
-  flex-wrap: wrap;
-  margin-top: 50px;
- margin-bottom: 50px;
-}
-
-ul {
-  list-style: none;
-}
-
-body{
-  margin: 1px 0;
-  background-color: #080808;
-  background-image: url('assets/Headerbg.png');
-  background-repeat: no-repeat; /* Prevent the image from repeating */
- 
-}
-.card-container {
- 
-  display: grid;
-  grid-template-columns: repeat(2, 1fr);
-  gap: 20px;
-  justify-items: center;
-}
-
-.container{
- 
-  position: relative;
-  margin-top: 49px;
-  margin-left: 69px;
-  height: 368px;
-  width: 579px;
-  border-radius: 10px;
-  box-shadow: 2px 2px 1px rgba(0,0,0,0.2);
-}
-.container:nth-child(1) {
-  background-image: linear-gradient(rgba(255, 132, 0, 0.5), rgba(255, 204, 0, 0.5)), url('assets/cloudbg.png');
-}
-.container:nth-child(2) {
-  background-image: linear-gradient(rgba(0, 42, 255, 0.5), rgba(0, 51, 255, 0.5)), url('assets/cloudbg.png');
-}
-
-/*  CONTENT  */
-.content { 
-  width: 579px;
-  height: 218px;
-  border-radius: 8px 8px 0 0;
-  background-size: cover;
-  background-repeat: no-repeat;
-  background-position: center;
-  position: relative;
-  margin-top: 0;
-}
-
-
-.sun{
-  z-index: 1000;
-  font-size: 15px !important;
-}
-
-.location{
-  position: absolute;
-  font-family: "Roboto", sans-serif;
-  font-weight: 100;
-  font-size: 20px;
-  width: 177px;
-  height: 26px;
-  left: 69px;  
-  top: 49px;
-}
-.dateAndTime{
-  position: absolute;
-  font-family: "Roboto", sans-serif;
-  font-weight: 30;
-  font-size: 12px;
-  width: 150px;
-  height: 16px;
-  left: 80px;  
-  top: 91px;
-}
-.description{
-  position: absolute;
-  font-family: "Roboto", sans-serif;
-  font-weight: 40;
-  font-size: 14px;
-  width: 103px;
-  height: 16px;
-  left: 132px;  
-  top: 149px;
-}
-.temp{
-  z-index: 1000;
-  width: 126px;
-  height: 48px;
-  position: absolute;
-  font-family: "Noto Sans", sans-serif;
-  font-size: 44px;
-  font-weight: 20;
-  right: 100px;
-  top: 48px;
-}
-.tempMax{
-  z-index: 1000;
-  width: 130px;
-  height: 40px;
-  position: absolute;
-  font-family: "Noto Sans", sans-serif;
-  font-size: 14px;
-  font-weight: 100;
-  right: 110px;
-  top: 150px;
-}
-.tempMin{
-  z-index: 1000;
-  width: 130px;
-  height: 40px;
-  position: absolute;
-  font-family: "Noto Sans", sans-serif;
-  font-size: 14px;
-  font-weight: 100;
-  right: 110px;
-  top: 110px;
-}
-.bottom{
-width: 579px;
-height: 151px;
-border-radius: 8px;
-background-color: #383b47;
-display: grid;
-justify-content: center;
-grid-template-columns: repeat(3, 160px);
-gap: 2px;
-border-radius: 0 0 10px 10px;
-position: relative;
-font-family: "Noto Sans", sans-serif;
-}
-.bottom > div {
-  list-style: none;
-  text-align: left;
-  padding: 20px 4px;
-  font-size: 14px;
-  justify-content: center;
-  margin: 5px;
-  height: 50%;
-  bottom: 5px;
-  top: 5px;
-}
-.vertical {
-  border-left: 1px solid rgb(134, 134, 149);
-  height: 100px;
-  position:relative;
-  margin-left: -6px;
-}
-.search-bar {
-  display: flex;
-  width: 25%;
-  align-items: center;
-  padding: 0px;
-  background-color: #0d1013;
-  border-radius: 5px;
-}
-.searchBarWrapper{
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  height: 50px;
-}
-button {
-  padding: 8px 16px;
-  margin-left: 0px;
-  background-color: #6b28a9;
-  color: #fff;
-  border: none;
-  border-radius: 4px;
-  cursor: pointer;
-  font-size: 14px;
-  outline: none;
-}
-input[type='text'] {
-  flex: 1;
-  padding: 8px;
-  border: none;
-  border-radius: 4px;
-  font-size: 14px;
-  outline: none;
-  background-color: #0d1013;
-}
-.footer {
-  background-color: #121111;
-  padding: 10px;
-  text-align: center;
-  
-}
-@media (max-width: 768px) {
-  .logo {
-    display: flex;
-    flex-direction: column;
-    position: relative;
-    align-items: center;
-    margin-top: 50px;
-    justify-content: center; 
-    width: 100%; 
-    height: 85px;
-  }
-  .app {
-    display: flex;
-    flex-direction: column;
-    display: block;
-    margin-top: 50px;
-   margin-bottom: 50px;
-  
-  }
-  .card {
-    width: 70%;
-    height: 50%;
-  }
-  
-  .container{
-    width: 100%;
-  }
-  .container:nth-child(1) {
-    background-image: linear-gradient(rgba(0, 179, 255, 0.5), rgba(0, 179, 255, 0.5)), url('assets/cloudbg.png');
-  }
-  .content {
-    align-items: center;
-    height: 200px;
-    margin-top: 0;
-  }
-  
-  .location{
-    font-weight: 100;
-    font-size: 20px;
-    left: 5%; 
-  }
-  .dateAndTime{
-    width: 103px;
-    height: 16px;
-    left: 6%;  
-  
-  }
-  .description{
-    font-weight: 20;
-    font-size: 13px;
-    left: 8%;  
-  }
-  .temp{
-    
-    font-size: 20px;
-    font-weight: 5;
-    right: 55%;
-    top: 30px;
-  }
-  .tempMinMax{
-
-    font-size: 18px;
-    font-weight: 5;
-    right: 50%;
-    top: 129px;
-  }
-  
-  .bottom > div {
-    padding: 5px 0;
-    font-size: 16px;
-    justify-content: center;
-    margin: 5px;
-    height: 20%;
-    bottom: 5px;
-    top: 5px;
-  }
-  .bottom {
-    width: 100%;
-    height: auto;
-    grid-template-columns: repeat(1, 1fr);
-    gap: 10px;
-  }
-
-  .search-bar {
-    display: flex;
-    width: 100%;
-    align-items: center;
-    padding: 0px;
-
-    border-radius: 5px;
-  }
-  .searchBarWrapper{
-    width: 100%;
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    height: 50px;
-  }
-  button {
-    padding: 8px 16px;
-    margin-left: 0px;
-    background-color: #6b28a9;
-    color: #fff;
-    border: none;
-    border-radius: 4px;
-    cursor: pointer;
-    font-size: 14px;
-    outline: none;
-  }
-  input[type='text'] {
-    flex: 1;
-    padding: 8px;
-    border: none;
-    border-radius: 4px;
-    font-size: 14px;
-    outline: none;
-    background-color: #0d1013;
-  }
-}
-
-
-
