+++ src/pages/CityWeather.css
@@ -1,18 +1,34 @@
-.app {
+*{
+    padding: 0;
+    margin: 0;
+    color: #FFF;
+  }
+
+  .body{
+    min-height: 100vh;
     display: flex;
     flex-wrap: wrap;
     margin-top: 50px;
-   margin-bottom: 50px;
-  }
-  
-  ul {
-    list-style: none;
-  }
-  
-  body{
+    margin-bottom: 50px;
     margin: 1px 0;
     background-color: #080808;
-    background-image: url('assets/Headerbg.png');
+    background-image: url('../assets/Headerbg.png');
     background-repeat: no-repeat; /* Prevent the image from repeating */
    
   }
+  @media screen and (max-width: 768px) {
+    .body {
+      min-height: 100vh;
+      display: flex;
+      flex-wrap: wrap;
+      margin-top: 0;
+      margin-bottom: 0;
+      margin: 0;
+      background-color: #080808;
+      background-image: url('../assets/Headerbg.png');
+      background-repeat: no-repeat; /* Prevent the image from repeating */
+      background-size: cover;
+      width: 100%;
+    }
+  }
+
