+++ src/pages/dashboard.css
@@ -1,0 +1,94 @@
+*{
+    padding: 0;
+    margin: 0;
+    color: #FFF;
+  }
+
+  .body{
+    min-height: 100vh;
+    display: flex;
+    flex-wrap: wrap;
+    margin-top: 50px;
+    margin-bottom: 50px;
+    margin: 1px 0;
+    background-color: #080808;
+    background-image: url('../assets/Headerbg.png');
+    background-repeat: no-repeat; /* Prevent the image from repeating */
+   
+  }
+
+  .search-bar {
+    display: flex;
+    width: 25%;
+    align-items: center;
+    padding: 0px;
+    border-radius: 5px;
+  }
+  .searchBarWrapper{
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    height: 50px;
+    width: 100%; 
+   
+  }
+  button {
+    padding: 8px 16px;
+    margin-left: 0px;
+    background-color: #6b28a9;
+    color: #fff;
+    border: none;
+    border-radius: 4px;
+    cursor: pointer;
+    font-size: 14px;
+    outline: none;
+  }
+  input[type='text'] {
+    flex: 1;
+    padding: 8px;
+    border: none;
+    border-radius: 4px;
+    font-size: 14px;
+    outline: none;
+    background-color: #0d1013;
+  }
+  .logo {
+    display: flex;
+    position: relative;
+    align-items: center;
+    margin-top: 50px;
+    justify-content: center; 
+    width: 100%; 
+    height: 85px;
+  }
+  .footer {
+    background-color: #121111;
+    padding: 10px;
+    text-align: center;
+    
+  }
+  @media (max-width: 768px) {
+    .body {
+      flex-direction: column;
+    }
+  
+    .search-bar {
+      width: 100%;
+    }
+  
+    .searchBarWrapper {
+      height: 100px;
+    }
+  
+    button {
+      width: 100%;
+    }
+  
+    input[type='text'] {
+      width: 100%;
+    }
+  
+    .logo {
+      height: 100px;
+    }
+  }
